+++ NEWS.md
@@ -6,7 +6,11 @@
 ### Enhancements
 
 * `DurgaDiff` takes a parameter `ef.size.cex` that controls the size of the effect size central tendency symbol.
-* `DurgBrackets` prevents bracket text from overlapping.
+* `DurgaBrackets` prevents bracket text from overlapping.
+
+### Bug fixes
+
+* `DurgaDiff` now calculates CI of group mean for groups with at least 3 values and 2 distinct values. Previously it was only calculated for at least 3 distinct values.
 
 ## Durga 2.0