+++ NEWS.md
@@ -6,10 +6,10 @@
 ### Enhancements
 
 * `DurgaDiff` takes a parameter `ef.size.cex` that controls the size of the effect size central tendency symbol.
-* `DurgaBrackets` prevents bracket text from overlapping.
 
 ### Bug fixes
 
+* `DurgaBrackets` prevents bracket text from overlapping.
 * `DurgaDiff` now calculates CI of group mean for groups with at least 3 values and 2 distinct values. Previously it was only calculated for at least 3 distinct values.
 * `DurgaDiff` now correctly handles `R = NA` as documented; does not calculate CIs