+++ NEWS.md
@@ -1,7 +1,7 @@
 # Changes to the `Durga` package
 
 
-## Durga 2.0.0.9000
+## Durga 2.1.0
 
 ### Enhancements